Auxiliary Implementations

OceanSonar.Auxiliary.bivariate — Method

function bivariate(
    x::AbstractVector{<:Real},
    yz::AbstractVector{<:AbstractMatrix{<:Real}}
)

Returns a bivariate linear interpolation and flat extrapolation of input vector x, associated with input row vectors yz[n][:, 1] which pairs with output vectors yz[n][:, 2] for $n = 1, 2, ...$.

OceanSonar.Auxiliary.bivariate — Method

bivariate(
    x::AbstractVector{<:Real}, y::AbstractVector{<:Real}, Z::AbstractMatrix{<:Real})

Returns a bivariate linear interpolation and flat extrapolation of input vectors x and y for output matrix Z.

OceanSonar.Auxiliary.bivariate — Method

function bivariate(fcn::Function; x_imp = [-Inf, Inf], y_imp = [-Inf, Inf])

Checks the inputted function is bivariate. Returns the function fcn and a vector of important input points for the domain of inputs (x, y) for fcn(x, y): extrema(x_imp) defines the x domain limits for x, extrema(y_imp) the domain limits for y.

OceanSonar.Auxiliary.bivariate — Method

bivariate(z::Real)

Returns a bivariate function of constant output z.

OceanSonar.Auxiliary.uniquesort! — Function

uniquesort!

Composition of sort! then unique.

OceanSonar.Auxiliary.univariate — Method

function univariate(x::AbstractVector{<:Real}, y::AbstractVector{<:Real})

Returns a univariate linear interpolation and flat extrapolation of input vector x for output vector y.

OceanSonar.Auxiliary.univariate — Method

function univariate(fcn::Function; x_imp = [-Inf, Inf])

Checks the inputted function is univariate. Returns the function fcn and a vector of important input points for the domain of input x for fcn(x): extrema(x_imp) defines the domain limits of fcn.

OceanSonar.Auxiliary.univariate — Method

function univariate(y::Real; x_imp = [-Inf, Inf])

Returns a univariate function of constant output y.